-changed library type to object library
