--- include/figcone_ini/detail/.gitignore
@@ -1,1 +1,0 @@
-external/
