+++ src/paramparser.cpp
@@ -1,20 +1,18 @@
-#ifndef FIGCONE_INI_PARAMPARSER_H
-#define FIGCONE_INI_PARAMPARSER_H
-
+#include "paramparser.h"
 #include "utils.h"
 #include "stream.h"
 #include <figcone_tree/tree.h>
 #include <figcone_tree/errors.h>
 #include <utility>
-#include <optional>
 #include <sstream>
 
 namespace figcone::ini::detail {
 
-inline std::optional<std::string> readParam(Stream& stream,
-                                   const std::string& wordSeparator,
-                                   const std::vector<std::string>& paramListValue,
-                                   const std::string& paramName)
+std::optional<std::string> readParam(
+        Stream& stream,
+        const std::string& wordSeparator,
+        const std::vector<std::string>& paramListValue,
+        const std::string& paramName)
 {
     auto quotedParam = readQuotedString(stream);
     if (quotedParam)
@@ -31,7 +29,9 @@ inline std::optional<std::string> readParam(Stream& stream,
     }
 }
 
-inline std::optional<std::vector<std::string>> readParamList(const std::string& paramName, const std::string& paramValue)
+std::optional<std::vector<std::string>> readParamList(
+        const std::string& paramName,
+        const std::string& paramValue)
 {
     auto inputStream = std::stringstream{paramValue};
     auto stream = Stream{inputStream};
@@ -54,8 +54,7 @@ inline std::optional<std::vector<std::string>> readParamList(const std::string& 
             skipWhitespace(stream, true);
             if (stream.peek() == endOfList || stream.atEnd())
                 throw ConfigError{"Parameter list '" + paramName + "' element is missing"};
-        }
-        else if (stream.peek() == endOfList) {
+        } else if (stream.peek() == endOfList) {
             stream.skip(1);
             return paramValueList;
         }
@@ -63,7 +62,7 @@ inline std::optional<std::vector<std::string>> readParamList(const std::string& 
     throw ConfigError{"Wrong parameter list '" + paramName + "' format: missing ']' at the end"};
 }
 
-inline std::string readParam(const std::string& paramValue)
+std::string readParam(const std::string& paramValue)
 {
     auto inputStream = std::stringstream{paramValue};
     auto stream = Stream{inputStream};
@@ -76,5 +75,3 @@ inline std::string readParam(const std::string& paramValue)
 }
 
 }
-
-#endif //FIGCONE_INI_PARAMPARSER_H
